Api files moved to webapi package Documentation for query parameters methods
